BOL_Extraction/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Data Handling & Image Processing
pandas
numpy
Pillow
PyMuPDF
opencv-python-headless

# AI & Machine Learning
# Specific versions as requested
paddleocr==2.10.0
paddlepaddle==3.0.0

# Transformer models for table detection
torch
transformers
# Required for table detection models
timm

# HTTP Requests
requests

# Other utilities
python-multipart